ApunableBot.java: Add delete feature and modularize the code
User can input "delete <index>" to delete the task.

The switch in main function is too long, make it difficult to add new
features, need to abstract away some part of it.

Add helper functions to abstract away the essential part of switch, can
shorten the code and make it easier to maintain.