Update user guide and fix some minor bugs
There is no help command, and the way to use chatbot is not that
self-explanatory. It is hard for the user to know the type of commands
offered and their usage.

Let's update the user guide to include the features offered and the way
to use the chatbot. UG from AB3 is refered when creating this guide.

AB3 is an existing command line application with comprehensive UG. It
has been through many iterations and has include almost all the
necessary information that a new user need to know.

Besides, some small changes on `ContactEdit` and `ContactDelete` to save
contacts into hard disk after the commands.